Improvement on detecting outline incompatibilities
The outlines used for test 4,9, "incompatible constructions for glyph" were taken from the designspace. But as these are MathGlyphs with added on-curve bcps, actual differences in bcp structures in the ufo went undetected.
This reads the glyphs from the master UFOs and then tests. This will improve detecting issues that are raised by fontmake/.

Changes to the test fonts: new glyph with offcurve differences.
Add familynames to the test fonts so we can see which one we have open in the editor.
Normaliser differences.