Fixes related to assessing kernin
Be more precise what we test in kerning dicts. This may not have been the issue, but it's better to be explicit.
Problem 5,0 was reported wrong in systems in which one master did not have kerning, it would point to the default as the one missing kerning. Also address the same issue when reporting problem 5,6.