Rephrase Feature error message and provide more details
When checking the features, designspaceproblems tries to compile the features with FontTools FeaLib. This adds the error message to the Problem.details so we might learn something from this downstream.

The error message is rephrased into "problem parsing the feature file". File corruption is a different problem. We can read the file, but the parser is confused about it.

Separate question: is it useful to check the features as part of the designspaceproblems. One example at hand shows that FeaLib triggers an exception in data that otherwise compiles into a font.